+++ CHANGELOG.md
@@ -9,11 +9,15 @@
 * delay judging last word
 * not refresh graph while typing
 * upload texts mechanism (submit/accept)
-* bootstrap 5.3.8 widh CDN
+* bootstrap 5.3.8 - does not go well
+* delete old typing data
+  delete from results where timestamp < '2026-03-01';
+* (reset) in dev namespace?
+
 
 ## 5.2.4 (2026-03-14)
 
-- bootstrap again
+- **changed** after login, go to `/todays`
 
 ## 5.2.3 (2026-03-14)